requirements.txt
tensorflow
flask
numpy
pillow
scipy
matplotlib
scikit-learn
gunicorn
